feat: create debug controller with dto
